dqbuf_is_cc_continuous bug fixes
In the situation where the queue size was 5, and a buffer with a stale
USB CC value was received, the buffer would be popped, and the queue
size would drop to 4. Thus, even though the CCs were continuous,
subsequent calls to confirm if the CCs are continuous were failing due
to the queue size requirement not being met. This caused incorrect
timeout errors.

Also fixed a separate bug where the buffer top pointer wouldn't get
updated if the queue size became just 1 after an stale buffer was
popped.